requirements.txt
aioredis==1.3.1
alembic==1.4.3
async-timeout==3.0.1
attrs==20.2.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
hiredis==1.1.0
idna==2.10
importlib-metadata==2.0.0
iniconfig==1.0.1
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.3
MarkupSafe==1.1.1
more-itertools==8.5.0
packaging==20.4
pluggy==0.13.1
psycopg2==2.8.6
py==1.9.0
pyparsing==2.4.7
pytest==6.0.2
python-dateutil==2.8.1
python-dotenv==0.14.0
python-editor==1.0.4
redis==3.5.3
requests==2.24.0
six==1.15.0
SQLAlchemy==1.3.19
toml==0.10.1
urllib3==1.25.10
Werkzeug==1.0.1
WTForms==2.3.3
zipp==3.2.0
sanic==21.3.4
celery==5.0.5